Inline logical vector views into vector module
